Vid-OCR/requirements.txt
beeply==1.0.2
imutils==0.5.2
numpy==1.16.4
opencv-python==4.1.0.25
Pillow==6.0.0
playsound==1.2.2
pytesseract==0.2.6
tesseract==0.1.3
